+++ src/rest/types/MapperFile.ts
@@ -1,0 +1,2 @@
+
+export type MapperFile = { id: string, displayName: string };