+++ src/rest/types/RestTypes.ts
@@ -1,0 +1,2 @@
+export type { MapperFile } from "./MapperFile";
+export type { MapperUpdate, MapperVersion, MapperUpdateRequest} from "./MapperUpdate";